backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
python-multipart>=0.0.9
email-validator>=2.1.0
requests>=2.31.0
Pillow>=10.0.0
